Crate nexus

Bindings for Raidcore Nexus addons.

Usage

use nexus::{
    gui::{register_render, render, RenderType},
    imgui::Window
};

nexus::export! {
    name: "My Addon",
    signature: -0x12345678,
    load: || {
        register_render(RenderType::Render, render!(|ui| {
            Window::new("My Window").build(ui, || {
                ui.text("Hello World");
            });
        }));
    }
}

Re-exports

pub use self::addon::AddonFlags;

pub use self::addon::AddonLoad;

pub use self::addon::AddonUnload;

pub use self::addon::UpdateProvider;

pub use imgui;

Modules

addon

Addon information.

alert

Alert notifications displayed to the user.

data_link

Data link for sharing resources.

event

Event system.

font

Font loading.

gamebind

Game keybinds.

gui

ImGui rendering via imgui-rs.

hook

Hooking via MinHook.

keybind

Addon keybinds.

localization

Localization of strings.

log

Logging.

paths

Paths.

quick_access

Quick access creation.

rtapi

Bindings for Nexus RealTime API.

texture

Texture loading.

updater

Update management.

v2

Addon API version 2.

v3

Addon API version 3.

v4

Addon API version 4.

v6

Addon API version 6.

wnd_proc

Windows WNDPROC.

Macros

event_consume

Macro to wrap an event callback.

event_subscribe

Macro to subscribe to an event with a wrapped callback.

export

Creates addon exports for Raidcore Nexus.

font_receive

Macro to wrap a font receive callback.

keybind_handler

Macro to wrap a keybind handler callback.

render

Macro to wrap an ImGui render callback.

texture_receive

Macro to wrap a texture receive callback.

Structs

AddonApi

Nexus addon API (version 6).

DataLinkApi

EventApi

FontApi

GameBindApi

InputBindsApi

LocalizationApi

MinHookApi

PathApi

QuickAccessApi

RendererApi

Revertible

A revertible action.

SupportedFields

Fields supported by the export macro.

TextureApi

UiApi

WndProcApi

Functions

addon_apiDeprecated

Returns the Nexus AddonApi instance.

on_unload

Adds a new action to be perform on unload.

ui⚠

Returns an [imgui::Ui] for rendering a frame.