nexus

Module font

Font loading.

Macros

font_receive

Macro to wrap a font receive callback.

Functions

add_font_from_file

Adds a font from a file path and sends updates to the callback.

add_font_from_memory

Adds a font from memory and sends updates to the callback.

add_font_from_resource

Adds a font from a resource and sends updates to the callback.

get_font

Registers a new callback to receive the font with the given identifier.

release_font

Releases a previously registered callback for the font with the given identifier.

resize_font

Resizes an existing font, sending the update to registered callbacks.

Type Aliases

RawFontAddFromFile

RawFontAddFromMemory

RawFontAddFromResource

RawFontGet

RawFontReceive

RawFontRelease

RawFontResize