nexus
Module gui

ImGui rendering via imgui-rs.

Macros

render

Macro to wrap an ImGui render callback.

Structs

RenderTypeIter

An iterator over the variants of RenderType

Enums

RenderType

Type of render callback.

Constants

IMGUI_VERSION

ImGui version.

Functions

deregister_close_on_escape

Deregisters a window name to listen to on escape.

register_close_on_escape

Registers a window name to get its bool toggled when escape is pressed.

register_render

Registers a new ImGui render callback of the given RenderType.

unregister_render

Unregisters a previously registered ImGui render callback.

Type Aliases

ImguiFree

ImguiMalloc

RawGuiAddRender

RawGuiDeregisterCloseOnEscape

RawGuiRegisterCloseOnEscape

RawGuiRemRender

RawGuiRender