nexus
Module localization

Localization of strings.

Functions

set_translation

Attempts to set a translated string for the given identifier and language at runtime.

translate

Attempts to translate the identifier into the current active language. Returns the same identifier if not available.

translate_to

Attempts to translate the identifier into the given language. Returns the same identifier if not available.

Type Aliases

RawLocalizationSet

RawLocalizationTranslate

RawLocalizationTranslateTo