nexus
Module quick_access

Quick access creation.

Functions

add_quick_access

Adds a new shortcut icon to the quick access with the given texture identifiers. When clicked the given keybind identifier is triggered.

add_quick_access_context_menu

Adds a new RawGuiRender callback for the shortcut context menu.

add_shortcutDeprecated

Adds a new shortcut icon to the quick access with the given texture identifiers. When clicked the given keybind identifier is triggered.

add_simple_shortcutDeprecated

Adds a new RawGuiRender callback fired when the quick access is right-clicked.

notify_quick_access

Sends a notification to the given quick access shortcut.

remove_quick_access

Removes a previously registered shortcut from the quick access.

remove_quick_access_context_menu

Removes a previously registered shortcut context menu callback.

remove_shortcutDeprecated

Removes a previously registered shortcut from the quick access.

remove_simple_shortcutDeprecated

Removes a previously registered simple shortcut callback.

Type Aliases

RawQuickAccessAddContextMenu

RawQuickAccessAddContextMenu2

RawQuickAccessAddShortcut

RawQuickAccessGeneric