nexus::rtapi

Struct RealTimeApi

pub struct RealTimeApi(/* private fields */);

Interface for RealTime API.

Implementations

impl RealTimeApi

pub const SIG: i32 = 620_863_532i32

Signature of the RealTime API addon.

pub fn get() -> Option<Self>

Retrieves the RealTime API interface.

pub const fn as_ptr(&self) -> *const RealTimeData

Returns a raw pointer to the RealTimeData.

pub const fn as_non_null(&self) -> NonNull<RealTimeData>

Returns a NonNull to the RealTimeData.

pub unsafe fn as_ref(&self) -> &RealTimeData

Returns a reference to the value.

This is dangerous since the memory is volatile and holding a reference can easily violate aliasing rules.

Safety

See NonNull::as_ref.

pub fn is_active(&self) -> bool

Checks whether the link is active.

pub fn read_game(&self) -> Option<GameData>

Reads the GameData.

pub fn read_world(&self) -> Option<WorldData>

Reads the WorldData.

pub fn read_group(&self) -> Option<GroupData>

Reads the GroupData.

pub fn read_player(&self) -> Option<PlayerData>

Reads the PlayerData.

pub fn read_camera(&self) -> Option<CameraData>

Reads the CameraData.