nexus::v6

Struct AddonApi

#[repr(C)]
pub struct AddonApi {

    pub swap_chain: IDXGISwapChain,
    pub imgui_context: *mut ImGuiContext,
    pub imgui_malloc: Option<ImguiMalloc>,
    pub imgui_free: Option<ImguiFree>,
    pub renderer: RendererApi,
    pub request_update: RawRequestUpdate,
    pub log: RawLog,
    pub ui: UiApi,
    pub path: PathApi,
    pub min_hook: MinHookApi,
    pub event: EventApi,
    pub wnd_proc: WndProcApi,
    pub input_binds: InputBindsApi,
    pub game_bind: GameBindApi,
    pub data_link: DataLinkApi,
    pub texture: TextureApi,
    pub quick_access: QuickAccessApi,
    pub localization: LocalizationApi,
    pub font: FontApi,

}

Nexus addon API (version 6).

Fields
swap_chain: IDXGISwapChain

DirectX swap chain.

imgui_context: *mut ImGuiContext

ImGui context.

imgui_malloc: Option<ImguiMalloc>

ImGui malloc function.

imgui_free: Option<ImguiFree>

ImGui free function.

renderer: RendererApi

Rendering API.

request_update: RawRequestUpdate

Downloads the addon available at remote without checking its version.

log: RawLog

Logs a message to the log window and log file.

Supports custom coloring for addon window messages, for example <c=#FF0000>this text is red</c>.

ui: UiApi

Ui API.

path: PathApi

Paths API.

min_hook: MinHookApi

MinHook API.

event: EventApi

Event API.

wnd_proc: WndProcApi

WNDPROC API.

input_binds: InputBindsApi

Input keybinds API.

game_bind: GameBindApi

Game keybinds API.

data_link: DataLinkApi

Data Link API.

texture: TextureApi

Texture Api.

quick_access: QuickAccessApi

Quick Access API.

localization: LocalizationApi

Localization API.

font: FontApi

Font API.

Implementations

impl AddonApi

pub const VERSION: i32 = 6i32

Nexus Addon API version.

pub fn get_d3d11_device(&self) -> Option<ID3D11Device>

Retrieves the DirectX 11 device associated with the swap chain.

impl AddonApi

pub fn get() -> &'static Self

Returns the Nexus AddonApi instance.

Panics if called before initialization.