Macro nexus::event_subscribe

macro_rules! event_subscribe {
    ( unsafe $event:expr , $ty:ty , $callback:expr $(,)? ) => { ... };
    ( unsafe $event:expr => $ty:ty , $callback:expr $(,)? ) => { ... };
    ( $event:expr , $ty:ty , $callback:expr $(,)? ) => { ... };
    ( $event:expr => $ty:ty , $callback:expr $(,)? ) => { ... };
}

Macro to subscribe to an event with a wrapped callback.

This macro is unsafe. See event_subscribe_typed for more information.

Returns a Revertible to revert the subscribe.

Usage

unsafe {
    event_subscribe!("MY_EVENT" => i32, |args| {
        use nexus::log::{log, LogLevel};
        log(LogLevel::Info, "My Addon", format!("Received MY_EVENT with {args:?}"));
    })
}.revert_on_unload();

The event identifier may be dynamic and the callback may be a function name.

let event: &str = "MY_EVENT";
fn event_callback(event_args: Option<&i32>) {
    use nexus::log::{log, LogLevel};
    log(LogLevel::Info, "My Addon", format!("Received MY_EVENT with {event_args:?}"));
}
let revertible = unsafe { event_subscribe!(event => i32, event_callback) };
revertible.revert();

The unsafe keyword may be moved into the macro call:

event_subscribe!(unsafe "MY_EVENT" => (), event_callback);

Safety

See event_subscribe_typed.