nexus
Module texture

Texture loading.

Macros

texture_receive

Macro to wrap a texture receive callback.

Structs

Texture

A loaded texture.

Functions

get_texture

Attempts to retrieve a texture by its identifier.

get_texture_or_create_from_file

Attempts to retrieve a texture or creates it from the given file path.

get_texture_or_create_from_memory

Attempts to retrieve a texture or creates it from the given memory.

get_texture_or_create_from_resource

Attempts to retrieve a texture or creates it from the given resource.

get_texture_or_create_from_url

Attempts to retrieve a texture or creates it from the given URL.

load_texture_from_file

Loads a texture from the given file path.

load_texture_from_memory

Loads a texture from the given memory. /// You can create a RawTextureReceiveCallback using the texture_receive macro.

load_texture_from_resource

Loads a texture from the given resource.

load_texture_from_url

Loads a texture from the given URL.

Type Aliases

RawTextureGet

RawTextureGetOrCreateFromFile

RawTextureGetOrCreateFromMemory

RawTextureGetOrCreateFromResource

RawTextureGetOrCreateFromUrl

RawTextureLoadFromFile

RawTextureLoadFromMemory

RawTextureLoadFromResource

RawTextureLoadFromUrl

RawTextureReceiveCallback